send(), sendto(), no longer raise errors if they fail to send the required number of bytes and return a count of the number of bytes transferred.
Check for -ve transfer counts in recv and recvfrom().

Fixed some typos and formatting issues.

Changed seterror() to take a fallback error string in case ici_buf
cannot be resized.  The string is used as the error string if no
formatted string is required.  This needs to looked at to reduce
string duplication.

Removed all tabs.